Fix crash on screen orientation change
When screen is changed from Portrait to landscape or vice-versa and then country row is tapped to display the details, the app would crash since repository is wiped out on orientation change but view model stays. This commit fixes the problem.
