+++ app/src/main/java/com/kodeco/android/countryinfo/MainActivity.kt
@@ -3,35 +3,17 @@ package com.kodeco.android.countryinfo
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.kodeco.android.countryinfo.api.CountryAPIService
 import com.kodeco.android.countryinfo.nav.CountryInfoNavHost
-import com.kodeco.android.countryinfo.repositories.CountryRepositoryImpl
 import com.kodeco.android.countryinfo.ui.theme.MyApplicationTheme
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
 
-private const val BASE_URL = "https://restcountries.com/v3.1/"
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        val moshi = Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            .build()
-
-        val retrofit = Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(MoshiConverterFactory.create(moshi))
-            .build()
-
-        val apiService: CountryAPIService = retrofit.create(CountryAPIService::class.java)
-
         setContent {
             MyApplicationTheme {
-                CountryInfoNavHost(repository = CountryRepositoryImpl(apiService))
+                CountryInfoNavHost(repository = (application as CountryInfoApp).repository)
             }
         }
     }
